Decoupling of grading roles (ie classwork, homework and exams) and the activities (approaches) that generate these grades. The activities need to be plugins. The alternative activities should not be hardcoded into the grading roles.
One possibility is delegation of eg, classwork_total to a classwork accessor, an instance of a Classwork class. The classwork accessor delegates to an Approach accessor which calls eg, either a CompComp class's methods or a Groupwork class's methods.

This will be the first commit.

The problem with this solution is the CompComp and Groupwork methods need the League object.